Debug all Test Classes

+++ src/test/java/CustomArrayListTestMe.java
@@ -141,7 +141,7 @@ class CustomArrayListTestMe<T> {
         @Test
         void should_remove_item_from_end_of_list() {
             Integer originalSize = sut.getSize();
-            assertEquals(originalSize - 1, sut.get(originalSize - 1));
+            assertEquals((Integer) (originalSize - 1), sut.get(originalSize - 1));
             sut.remove(originalSize - 1);
             assertEquals(originalSize - 1, sut.getSize());
         }
